Don't return edge list in contractor but modify graph in-place
